[refactor] rename LectureSWManagement to RegistrationSW
- Renamed to show more explicit names
