+++ src/components/tabs/RegistrationSWTab/RegistrationSWTab.stories.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import LectureSWManagementTab from '.';
+import RegistrationSWTab from '.';
 import { generateSWForLecture, generateStringArray } from '../../../__mocks__/create-mock';
 
 export default {
-  title: 'Container/LectureSWManagementTab',
-  component: LectureSWManagementTab,
+  title: 'Container/RegistrationSWTab',
+  component: RegistrationSWTab,
 };
 
 const manufacturings = generateStringArray(3);
 const items = [generateSWForLecture(), generateSWForLecture(), generateSWForLecture()];
-export const lectureSWManagementTab = () => <LectureSWManagementTab items={items} manufacturings={manufacturings} />;
+export const registrationSWTab = () => <RegistrationSWTab items={items} manufacturings={manufacturings} />;
