+++ src/components/tabs/RegistrationSWTab/styled.ts
